+++ packages/react/__tests__/FederatedComponent.spec.tsx
@@ -4,18 +4,18 @@ import '@testing-library/jest-dom';
 import { act, render, screen } from '@testing-library/react';
 import React from 'react';
 import FederatedComponent from '../src/components/FederatedComponent';
-import * as useRegistry from '../src/hooks/useRegistry';
-import index from './fixtures/appshell.index';
+import * as useGlobalConfig from '../src/hooks/useGlobalConfig';
 import manifest from './fixtures/Manifest';
 import TestComponent from './fixtures/TestComponent';
+import config from './fixtures/TestGlobalConfig';
 
-jest.mock('../src/hooks/useRegistry');
+jest.mock('../src/hooks/useGlobalConfig');
 
 const TestFallback = () => <div>loading...</div>;
 
 describe('FederatedComponent', () => {
   it('should match snapshot', async () => {
-    jest.spyOn(useRegistry, 'default').mockReturnValue(index);
+    jest.spyOn(useGlobalConfig, 'default').mockReturnValue(config);
     jest.spyOn(remoteLoader, 'default').mockReturnValueOnce(async () => [TestComponent, manifest]);
 
     const { container, findByText } = await act(() =>
@@ -28,18 +28,18 @@ describe('FederatedComponent', () => {
   });
 
   it('should render fallback while resource is pending', async () => {
-    jest.spyOn(useRegistry, 'default').mockReturnValue(index);
+    jest.spyOn(useGlobalConfig, 'default').mockReturnValue(config);
     jest.spyOn(remoteLoader, 'default').mockReturnValueOnce(() => [null, null]);
 
     await act(() =>
       render(<FederatedComponent remote="TestModule/TestComponent" fallback={<TestFallback />} />),
     );
 
-    await expect(screen.getByText(/loading/i)).toBeInTheDocument();
+    expect(screen.getByText(/loading/i)).toBeInTheDocument();
   });
 
   it('should render error when resource fails', async () => {
-    jest.spyOn(useRegistry, 'default').mockReturnValue(index);
+    jest.spyOn(useGlobalConfig, 'default').mockReturnValue(config);
     jest
       .spyOn(remoteLoader, 'default')
       .mockImplementationOnce(() => new Error('Failed to get resource'));
@@ -48,6 +48,6 @@ describe('FederatedComponent', () => {
       render(<FederatedComponent remote="TestModule/TestComponent" fallback={<TestFallback />} />),
     );
 
-    await expect(screen.getByText(/Error loading federated component/i)).toBeInTheDocument();
+    expect(screen.getByText(/Error loading federated component/i)).toBeInTheDocument();
   });
 });